Fixed some bugs & translations

+++ src/src/Controller/OrderController.php
@@ -77,8 +77,10 @@ class OrderController extends AbstractController
     public function delete(Request $request, Order $order, OrderRepository $orderRepository): Response
     {
         $this->denyAccessUnlessGranted('delete', $order);
-        if ($this->isCsrfTokenValid('delete' . $order->getId(), $request->request->get('_token'))) {
-            $orderRepository->remove($order, true);
+        if ($order->getStatus() == "Registered") {
+            if ($this->isCsrfTokenValid('delete' . $order->getId(), $request->request->get('_token'))) {
+                $orderRepository->remove($order, true);
+            }
         }
 
         return $this->redirectToRoute('app_order_index', [], Response::HTTP_SEE_OTHER);
